NeroRunner Revamp (#70)
* Nero Runner Revamp - Initial Commit

This commit adds in the revamped runner flows for both one time Prefix launches and shortcut launches.

* rename NeroSetting to CombinedSetting

makes it's purpose clearer and avoids ambiguity in case this is reused
elsewhere.

revert case back to NeroConstant, remove Scaling enum

move nvApi into map

this was done already in prefixRunner; syncing them up here.

* fix regressions from runner redo

Gamescope Switch Case:
- using { in c++ switches causes them to not flow down to the next case,
functioning as a soft break. This commit changes it to instead
encapsulate both scenarios, so that it can properly "fall through".
Other:
- fix arguments variable not being updated from method

fix Gamemode setting

We need to check the bool, not if it has a setting

fix formatting

This commit makes several small changes to align the application with
more consistent formatting to what's seen in other classes.

refactor wined3d logic

This was just really messy in general; I was following how the original
flow worked, but on my nth recheck I realized the flow could be
substantially simplified.

fix noD8VK logic

fix dll overriding

* revert logging flow

I really don't understand why returning the file and re-opening it
breaks the fabric of reality for the logfile, so this is staying as-is
for now.

formatting fixes

* Refactor Gamescope & Mangohud Logic

- add initSetting function for returning correct type of Setting
* used primarily in generic functions as a quick way of checking,
instead of setting the value
* CombinedSetting can't skip this value as occasionally its value is
required too.
refactor gamescope functions
- split Gamescope Windowed & Fullscreen logic into second function
* this function creates its own list which is prepended to the exist
arguments list to make logic clearer
- Moved Mangohud into seperate reusable method

* add helper struct for gamescope resolution

Contrary to what I thought, QMaps sort themselves automatically in
alphabetical order based off key.

This was resulting in the arguments coming in the opposite order they
would normally be input, ie "-h 1080 -w 1920", and meant that to keep
ordering consistent with expected order, we were iterating through the
QMap backwards.

This appears to have caused an issue with fullscreen gamescope, so this
commit changes it to be a temp struct that holds this information
instead.

* switch fpsLimit argument order

* change to skip empty strings

* add windowed case to gamescope scaling types

the windowed section didn't contain any unique logic, but we still need
to set the resolution options here and blank the windowArg.

* disable scaler option on backend

this will still have the frontend UI enabled, but with a warning that it
does nothing.

* fix shortcut command arguments